requirements.txt
fastapi
uvicorn[standard]
tensorflow>=2.15
numpy
pandas
pillow
matplotlib
scikit-learn
pyyaml
python-multipart
pytest
httpx
